requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.6
django-crispy-forms==1.9.1
django-filter==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
Markdown==3.2.2
newsapi-python==0.2.6
numpy==1.18.4
pandas==1.0.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
